Update RecentCrawlAdapter to move validation responsibility inside adapter rather than inject to test behaviour
